refactor(autofixer-smarthr-ui-migration): smarthrUiAlias関連の処理を共通化
helpers.jsを新規作成し、以下の機能を共通化:
- setupSmarthrUiAliasOptions: validSources拡張とaliasファイル判定
- isFileMatchingSmarthrUiAlias: ファイルパスマッチング

これにより各versionファイルで約30行の重複コードを削減。

変更内容:
- helpers.js を新規作成（migrator専用のヘルパー関数）
- v90-to-v91/index.js: 重複コードを削除し、helpers.jsを利用
- v91-to-v92/index.js: 重複コードを削除し、helpers.jsを利用
- DEVELOPER.md: 共通ヘルパーの使用方法と将来的な共通化方針を追記
- REFERENCE.md: 共通化の状況と今後の方針を追記

読みやすさ重視の方針により、チェッカー生成ヘルパーなど
過度な抽象化は見送り。v93以降でパターンが確立されたら再検討。
